Move EE MCP feature under chat
