+++ packages/web/src/ee/features/chat/mcp/actions.ts
