+++ packages/web/src/ee/features/chat/mcp/analytics.ts
