+++ packages/web/src/ee/features/chat/mcp/components/connectorRowInfo.tsx
