+++ packages/web/src/ee/features/chat/mcp/components/connectorToolDisclosure.test.tsx
@@ -1,7 +1,7 @@
 import { afterEach, describe, expect, test, vi } from 'vitest';
 import { cleanup, fireEvent, render, screen } from '@testing-library/react';
 import { ConnectorToolList, ConnectorToolTrigger } from './connectorToolDisclosure';
-import type { ServerToolsEntry } from '@/ee/features/mcp/types';
+import type { ServerToolsEntry } from '@/ee/features/chat/mcp/types';
 
 afterEach(() => {
     cleanup();
