+++ packages/web/src/ee/features/chat/mcp/components/mcpFavicon.tsx
