+++ packages/web/src/ee/features/chat/mcp/connectionStatus.test.ts
