+++ packages/web/src/ee/features/chat/mcp/dcrDiscovery.test.ts
