+++ packages/web/src/ee/features/chat/mcp/dcrDiscovery.ts
