+++ packages/web/src/ee/features/chat/mcp/errors.ts
