+++ packages/web/src/ee/features/chat/mcp/externalMcpError.test.ts
