+++ packages/web/src/ee/features/chat/mcp/externalMcpError.ts
