+++ packages/web/src/ee/features/chat/mcp/mcpToolMetadata.test.ts
