+++ packages/web/src/ee/features/chat/mcp/mcpToolMetadata.ts
