+++ packages/web/src/ee/features/chat/mcp/mcpToolRegistry.test.ts
