+++ packages/web/src/ee/features/chat/mcp/mcpToolRegistry.ts
