+++ packages/web/src/ee/features/chat/mcp/prefabMcpServers.test.ts
