+++ packages/web/src/ee/features/chat/mcp/types.ts
