+++ packages/web/src/ee/features/chat/mcp/utils.test.ts
