+++ packages/web/src/ee/features/chat/mcp/utils.ts
