+++ packages/web/src/features/chat/components/chatThread/tools/mcpToolComponent.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { CopyIconButton } from "@/app/(app)/components/copyIconButton";
-import { McpFavicon } from "@/ee/features/mcp/components/mcpFavicon";
+import { McpFavicon } from "@/ee/features/chat/mcp/components/mcpFavicon";
 import { useMcpServerIconMap } from "@/features/chat/mcpServerIconContext";
 import { cn } from "@/lib/utils";
 import { DynamicToolUIPart } from "ai";
